Add files via upload
Download XML formatted literature from tidyPMC using R script. If the download is successful, it will be updated to the database directory. If the download fails, the PMID will be saved to update-PMID_checked_PMC_download_fail.csv. Compare the PMID in the update_pmid.csv file with DB-PMID_status.csv, and output a list of PMIDs not included in this database. Use the R script to download XML formatted literature from tidyPMC. If the download is successful, it will be updated to the database directory. If the download fails, the PMID will be saved to update-PMID_checked_PMC_download_fail.csv. Use the R script to download the XML format literature from tidyPMC. If the download is successful, it will be updated to the database directory. If the download fails, the PMID will be saved to update-PMID_checked_PMC_download_fail.csv